requirements.txt
langchain==0.0.335
python-dotenv==1.0.1
langchain-openai==0.0.8